Add cordova device plugin and zoom in home page
The user can now zoom in/out the network graph in the home page with
two buttons.

Added the cordova device plugin to detect when the phone is ready and
the change the behavior of the backbutton on android.

Disabled the panel left swipe because of some scroll problems in
landscape mode.